pyproject.toml
[tool.poetry]
name = "rath"
version = "0.5.0"
description = "async transport-agnostic graphql client"
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
readme = "README.md"
packages = [{ include = "rath" }]

[tool.poetry.dependencies]
python = "^3.8"
koil = ">=1.0.0"
graphql-core = "^3.2.0"
pydantic = ">2"
websockets = { version = "^10.2", optional = true }
aiohttp = { version = "^3.8.2", optional = true }
certifi = { version = ">2021", optional = true }
httpx = { version = "^0.23.0", optional = true }
cryptography = { version = "^41.0.3", optional = true }

[tool.poetry.extras]
websockets = ["websockets"]
aiohttp = ["aiohttp", "certifi"]
httpx = ["httpx"]
signing = ["cryptography"]


[tool.poetry.group.dev.dependencies]
httpx = "^0.23.0"
autoflake = "^1.7.7"
pytest = "^7.2.0"
testcontainers = "^3.7.0"
pytest-qt = "^4.2.0"
pytest-asyncio = "^0.20.2"
aiohttp = "^3.8.3"
websockets = "^10.4"
black = "^22.10.0"
pytest-cov = "^4.0.0"
ruff = "^0.0.282"
cryptography = "^41.0.3"
pyjwt = "^2.8.0"
fakts = "1.0.0"
herre = "1.0.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ["venv/", "tests/", "examples/"]
ignore_missing_imports = true

[tool.ruff]
extend-select = ["ANN", "D1"]
extend-ignore = ["ANN101", "ANN002", "ANN003", "D100", "ANN401"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
    "examples",
]

# Same as Black.
line-length = 300


[[tool.pydoc-markdown.loaders]]
type = "python"
search_path = ["rath"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
qt_api = "pyqt5"
markers = [
    "integration: marks tests that require a running server",
    "qt: marks tests that require a running qt application",
    "public: marks tests that require a public api",
]

[tool.pydoc-markdown.renderer]
type = "docusaurus"
docs_base_path = "website/docs"
